Implement #391: Establish orchestration spine by decomposing main.rs
