+++ src/main.rs
@@ -7,7 +7,8 @@ use std::process::Command as ProcessCommand;
 use std::thread;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-use clap::{error::ErrorKind, Args, Parser, Subcommand, ValueEnum};
+use clap::ValueEnum;
+use cli::{CliLaneClaimSource, DisplayMode, ForgeStatusArg};
 use jade_symphony::agent::{
     backend_from_config, persist_prompt_artifact, usage_limit_pause_from_events, AgentBackend,
     ClaudeCodeBackend, CodexBackend, DryRunBackend, TmuxBackend, UsageLimitPause,
@@ -84,12 +85,13 @@ use jade_symphony::review::{
     transition_allowed_for_review_agent, write_review_job_ledger_record, FakeReviewBackend,
     FakeReviewOutcome, GeminiCliReviewBackend, GeminiReviewRecoveryPolicy, ReviewBackend,
     ReviewFreshnessInput, ReviewGateDecision, ReviewJob, ReviewJobState, ReviewOutcome,
-    ReviewRepeatedFailureEvidence, ReviewRequest, ReviewReworkClass, ReviewRunEligibility,
-    ReviewStaleReason,
+    ReviewRepeatedFailureEvidence, ReviewRequest, ReviewRunEligibility,
 };
+#[cfg(test)]
+use jade_symphony::review::{ReviewReworkClass, ReviewStaleReason};
 use jade_symphony::review_status::{
     load_review_status, render_project_inspect_review_summary, render_review_status_human,
-    ReviewStatusOptions, DEFAULT_RECENT_REVIEW_JOBS,
+    ReviewStatusOptions,
 };
 use jade_symphony::rework::rework_transition_expected;
 #[cfg(test)]
@@ -122,6 +124,8 @@ use jade_symphony::workspace::{
     Workspace,
 };
 use serde::Serialize;
+
+mod cli;
 
 const DEFAULT_RUN_LOOP_BASE_BRANCH: &str = "main";
 const DEFAULT_SESSION_STATUS_LINES: usize = 80;
@@ -14251,1444 +14255,6 @@ impl RunLoopOptions {
     }
 }
 
-impl Command {
-    fn parse(args: Vec<String>) -> Result<Self, String> {
-        if matches!(
-            args.first().map(String::as_str),
-            Some("help" | "--help" | "-h")
-        ) {
-            return Ok(Self::Help(usage()));
-        }
-
-        let argv = std::iter::once("jade-symphony".to_string())
-            .chain(args)
-            .collect::<Vec<_>>();
-        match Cli::try_parse_from(argv) {
-            Ok(cli) => Command::try_from(cli),
-            Err(error) if error.kind() == ErrorKind::DisplayHelp => {
-                Ok(Self::Help(error.to_string()))
-            }
-            Err(error) => Err(error.to_string()),
-        }
-    }
-}
-
-fn lane_command(lane: AgentSessionLaneArg, args: LaneCommandArgs) -> Result<Command, String> {
-    match args.command {
-        MainCommandArgs::Claim(claim) => Ok(Command::LaneClaim {
-            workflow_path: claim.workflow_path,
-            issue_ref: claim.issue_ref,
-            lane,
-            worker: claim.worker,
-            source: claim.source,
-            write: claim.write,
-        }),
-        MainCommandArgs::Once(args) if lane == AgentSessionLaneArg::Main => Ok(Command::RunOnce {
-            workflow_path: args.workflow_path,
-        }),
-        MainCommandArgs::Loop(args) if lane == AgentSessionLaneArg::Main => run_loop_command(args),
-        MainCommandArgs::Once(_) | MainCommandArgs::Loop(_) => {
-            Err("only the main lane supports once/loop through this command group".into())
-        }
-    }
-}
-
-#[derive(Debug, Parser)]
-#[command(
-    name = "jade-symphony",
-    about = "OpenAI Symphony-style orchestration harness with Jade Symphony extensions",
-    disable_help_subcommand = true,
-    arg_required_else_help = false
-)]
-struct Cli {
-    #[command(subcommand)]
-    command: Option<CliCommand>,
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: Option<PathBuf>,
-}
-
-#[derive(Debug, Subcommand)]
-enum CliCommand {
-    #[command(
-        next_help_heading = "Human / Operator operations",
-        alias = "plan-dispatch",
-        alias = "dry-run"
-    )]
-    Plan(WorkflowPathArgs),
-    #[command(alias = "validate-workflow")]
-    Validate(WorkflowPathArgs),
-    #[command(alias = "audit-project")]
-    Doctor(DoctorArgs),
-    #[command(name = "doctor-repair-human-review")]
-    DoctorRepairHumanReview(DoctorRepairArgs),
-    #[command(next_help_heading = "Human / Operator operations")]
-    Skills(SkillsArgs),
-    Profiles(WorkflowPathArgs),
-    Debug(WorkflowPathArgs),
-    #[command(
-        next_help_heading = "Lane orchestration",
-        name = "autopilot",
-        about = "Read-only all-lane planning preflight"
-    )]
-    Autopilot(AutopilotArgs),
-    Status(StatusArgs),
-    Clean(CleanArgs),
-    #[command(
-        next_help_heading = "Project / Agent internals",
-        about = "Discover and record per-issue git worktrees",
-        long_about = "Discover and record per-issue git worktrees.\n\n`workspace` is the safe local-worktree coordination surface for Main, Review, and Merge lanes. It discovers existing issue worktrees from the session registry, Main Workpad/timeline evidence, linked PR/branch hints, and `git worktree list`. It can ensure missing Review/Merge inspection worktrees under the configured workspace root, but it never runs `gh pr checkout`, switches branches, or changes the canonical repository checkout.\n\nUse `workspace show` before local Review or Merge inspection. Use `workspace adopt` only when an operator has selected an existing worktree that should become the canonical workspace evidence for the issue. Use `workspace ensure` only when no suitable candidate exists and local inspection is required."
-    )]
-    Workspace(WorkspaceArgs),
-    #[command(name = "session")]
-    Session(SessionArgs),
-    Project(ProjectArgs),
-    #[command(next_help_heading = "Lane orchestration", name = "main")]
-    Main(LaneCommandArgs),
-    #[command(name = "merge")]
-    Merge(MergeArgs),
-    Review(ReviewArgs),
-    #[command(name = "create-follow-up")]
-    CreateFollowUp(CreateFollowUpArgs),
-    #[command(next_help_heading = "Issue Forge")]
-    Forge(ForgeArgs),
-    #[command(
-        next_help_heading = "Reserved lifecycle topology",
-        about = "Reserved for future all-lane automatic orchestration"
-    )]
-    Run,
-    #[command(about = "Reserved for future Jade Symphony binary and skill upgrades")]
-    Upgrade,
-}
-
-#[derive(Debug, Args)]
-struct WorkflowPathArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    json: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct AutopilotArgs {
-    #[command(subcommand)]
-    command: AutopilotCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum AutopilotCommandArgs {
-    #[command(
-        about = "Plan Main, Review, and Merge lanes without mutating tracker or runtime state"
-    )]
-    Plan(AutopilotPlanArgs),
-}
-
-#[derive(Debug, Args)]
-struct AutopilotPlanArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    json: bool,
-}
-
-#[derive(Debug, Args)]
-struct ProjectStateArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long, value_enum, default_value_t = CliDisplayMode::Plain)]
-    display: CliDisplayMode,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct ProjectIssueArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    json: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct InspectArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long = "state")]
-    states: Vec<String>,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct DoctorRepairArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct DoctorArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: Option<PathBuf>,
-    #[arg(long)]
-    json: bool,
-    #[arg(long)]
-    strict: bool,
-    #[arg(long, value_enum, default_value_t = CliDisplayMode::Plain)]
-    display: CliDisplayMode,
-    #[arg(long)]
-    interactive: bool,
-    #[arg(long = "auto-fix")]
-    auto_fix: bool,
-    #[arg(long = "stale-after-ms", default_value_t = 10_800_000)]
-    stale_after_ms: u64,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    write: bool,
-    #[command(subcommand)]
-    action: Option<DoctorSubcommandArgs>,
-}
-
-#[derive(Debug, Args)]
-struct SkillsArgs {
-    #[command(subcommand)]
-    command: SkillsCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum SkillsCommandArgs {
-    #[command(about = "Report per-repo Jade Symphony skill readiness")]
-    Status(SkillsStatusArgs),
-}
-
-#[derive(Debug, Args)]
-struct SkillsStatusArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long = "suite-path")]
-    suite_path: Option<PathBuf>,
-    #[arg(long = "codex-dir")]
-    codex_dir: Option<PathBuf>,
-    #[arg(long = "gemini-dir")]
-    gemini_dir: Option<PathBuf>,
-    #[arg(long = "require-gemini")]
-    require_gemini: bool,
-    #[arg(long = "session-skills")]
-    session_skills: Vec<String>,
-    #[arg(long = "session-skills-file")]
-    session_skills_file: Option<PathBuf>,
-    #[arg(long)]
-    json: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct StatusArgs {
-    #[command(subcommand)]
-    command: StatusCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum StatusCommandArgs {
-    #[command(about = "Render the current runtime snapshot")]
-    Show(WorkflowPathArgs),
-    #[command(about = "Serve the current runtime snapshot once over loopback HTTP")]
-    Serve(StatusApiArgs),
-}
-
-#[derive(Debug, Subcommand)]
-enum DoctorSubcommandArgs {
-    Repair(DoctorRepairIssueArgs),
-}
-
-#[derive(Debug, Args)]
-struct DoctorRepairIssueArgs {
-    issue_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "move-need-human-input")]
-    move_need_human_input: bool,
-    #[arg(long = "mark-pr-ready")]
-    mark_pr_ready: bool,
-    #[arg(long = "confirm-handoff-ready")]
-    confirm_handoff_ready: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct StatusApiArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long, default_value = "127.0.0.1:8787")]
-    bind: SocketAddr,
-    #[arg(long)]
-    once: bool,
-}
-
-#[derive(Debug, Args)]
-struct RunLoopArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    max_iterations: Option<usize>,
-    #[arg(long)]
-    once: bool,
-    #[arg(long)]
-    write: bool,
-    #[arg(
-        long,
-        conflicts_with = "no_recover",
-        help = "Enable recover-first handling for interrupted Main runtime sessions (default in --write mode)"
-    )]
-    recover: bool,
-    #[arg(
-        long = "no-recover",
-        conflicts_with = "recover",
-        help = "Disable default recover-first handling in --write mode"
-    )]
-    no_recover: bool,
-    #[arg(long = "max-concurrent")]
-    max_concurrent: Option<usize>,
-    #[arg(long, value_enum, default_value_t = CliDisplayMode::Plain)]
-    display: CliDisplayMode,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DisplayMode {
-    Plain,
-    Tui,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
-enum CliDisplayMode {
-    Plain,
-    Tui,
-}
-
-impl From<CliDisplayMode> for DisplayMode {
-    fn from(value: CliDisplayMode) -> Self {
-        match value {
-            CliDisplayMode::Plain => Self::Plain,
-            CliDisplayMode::Tui => Self::Tui,
-        }
-    }
-}
-
-#[derive(Debug, Args)]
-struct CleanArgs {
-    #[command(subcommand)]
-    command: CleanCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum CleanCommand {
-    Plan(WorkflowPathArgs),
-    Audit(WorkflowPathArgs),
-}
-
-#[derive(Debug, Args)]
-struct CleanupWorkspacesArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct WorkspaceArgs {
-    #[command(subcommand)]
-    command: WorkspaceCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum WorkspaceCommandArgs {
-    #[command(
-        about = "List discovered issue worktrees and orphan hints",
-        long_about = "List discovered issue worktrees and orphan hints.\n\nThis is a read-only Project-wide inventory. It scans tracker issues, session registry records, Main Workpad/timeline evidence, linked PR/branch hints, and local `git worktree list` output. It reports candidates per issue and orphan-looking worktrees whose branch/path implies an issue not currently present in the fetched Project state."
-    )]
-    List(WorkspaceListArgs),
-    #[command(
-        about = "Show candidate worktrees for one issue",
-        long_about = "Show candidate worktrees for one issue.\n\nThis is the read-only preflight for Review and Merge agents before touching local files. It prints candidate worktrees, their branch/head metadata, evidence sources, warnings, and the canonical candidate when one can be chosen safely. Multiple strong candidates require operator choice through `workspace adopt` before local inspection should rely on a path."
-    )]
-    Show(WorkspaceShowArgs),
-    #[command(
-        about = "Record an operator-selected existing worktree",
-        long_about = "Record an operator-selected existing worktree as canonical workspace evidence for one issue.\n\n`workspace adopt` validates that the path is an existing git worktree for this repository and that its branch matches the issue/PR evidence. With `--write`, it writes a tracker workpad entry so later Main, Review, and Merge sessions can reuse the same workspace. It does not create a worktree, checkout a PR, switch branches, or mutate files in the selected worktree."
-    )]
-    Adopt(WorkspaceAdoptArgs),
-    #[command(
-        about = "Ensure a safe Review/Merge inspection worktree",
-        long_about = "Ensure a safe Review/Merge inspection worktree for one issue.\n\n`workspace ensure` first runs the same discovery as `workspace show` and reuses one suitable existing issue worktree when it can be chosen safely. If no suitable worktree exists, it prepares a git worktree only under the workflow-configured workspace root, using the linked PR branch or an explicit `--pr` / `--branch` argument. It never runs `gh pr checkout`, never switches the canonical checkout, refuses ambiguous candidates, and with `--write` records durable Workspace Evidence in the canonical issue workpad."
-    )]
-    Ensure(WorkspaceEnsureArgs),
-}
-
-#[derive(Debug, Args)]
-struct WorkspaceListArgs {
-    #[arg(
-        value_name = "path-to-WORKFLOW.md",
-        help = "Workflow config that defines the tracker, artifact roots, and workspace root"
-    )]
-    workflow_path: PathBuf,
-}
-
-#[derive(Debug, Args)]
-struct WorkspaceShowArgs {
-    #[arg(
-        value_name = "path-to-WORKFLOW.md",
-        help = "Workflow config that defines the tracker, artifact roots, and workspace root"
-    )]
-    workflow_path: PathBuf,
-    #[arg(help = "Issue identifier to inspect, for example #253")]
-    issue_ref: String,
-}
-
-#[derive(Debug, Args)]
-struct WorkspaceAdoptArgs {
-    #[arg(
-        value_name = "path-to-WORKFLOW.md",
-        help = "Workflow config that defines the tracker, artifact roots, and workspace root"
-    )]
-    workflow_path: PathBuf,
-    #[arg(help = "Issue identifier whose canonical workspace evidence should be updated")]
-    issue_ref: String,
-    #[arg(help = "Existing local git worktree path selected by the operator")]
-    path: PathBuf,
-    #[arg(
-        long,
-        help = "Write workspace adoption evidence to the tracker workpad"
-    )]
-    write: bool,
-    #[arg(
-        long = "dry-run",
-        help = "Preview adoption validation without writing tracker evidence"
-    )]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct WorkspaceEnsureArgs {
-    #[arg(
-        value_name = "path-to-WORKFLOW.md",
-        help = "Workflow config that defines the tracker, artifact roots, and workspace root"
-    )]
-    workflow_path: PathBuf,
-    #[arg(help = "Issue identifier whose Review/Merge inspection workspace should be ensured")]
-    issue_ref: String,
-    #[arg(
-        long = "pr",
-        help = "Optional PR number, URL, or ref to fetch when tracker-linked PR evidence is missing or ambiguous"
-    )]
-    pr_ref: Option<String>,
-    #[arg(
-        long,
-        help = "Optional branch/ref to use instead of the linked PR head branch"
-    )]
-    branch: Option<String>,
-    #[arg(
-        long,
-        help = "Create or reuse the worktree and write Workspace Evidence"
-    )]
-    write: bool,
-    #[arg(
-        long = "dry-run",
-        help = "Preview the reuse/create plan without creating worktrees or writing tracker evidence"
-    )]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct MergeOnceArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct MergeLoopArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    max_iterations: Option<usize>,
-    #[arg(long)]
-    once: bool,
-    #[arg(long)]
-    write: bool,
-    #[arg(
-        long,
-        conflicts_with = "no_recover",
-        help = "Enable recover-first handling for interrupted Merge loop claims (default in --write mode)"
-    )]
-    recover: bool,
-    #[arg(
-        long = "no-recover",
-        conflicts_with = "recover",
-        help = "Disable default recover-first handling in --write mode"
-    )]
-    no_recover: bool,
-    #[arg(long = "max-concurrent")]
-    max_concurrent: Option<usize>,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct LaneCommandArgs {
-    #[command(subcommand)]
-    command: MainCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum MainCommandArgs {
-    Claim(LaneClaimArgs),
-    Once(WorkflowPathArgs),
-    Loop(RunLoopArgs),
-}
-
-#[derive(Debug, Args)]
-struct LaneClaimArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    worker: String,
-    #[arg(long, value_enum, default_value_t = CliLaneClaimSource::Manual)]
-    source: CliLaneClaimSource,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
-enum CliLaneClaimSource {
-    Manual,
-    Loop,
-    Goal,
-}
-
-impl From<CliLaneClaimSource> for LaneClaimSource {
-    fn from(value: CliLaneClaimSource) -> Self {
-        match value {
-            CliLaneClaimSource::Manual => Self::Manual,
-            CliLaneClaimSource::Loop => Self::Loop,
-            CliLaneClaimSource::Goal => Self::Goal,
-        }
-    }
-}
-
-#[derive(Debug, Args)]
-struct AgentSessionArgs {
-    #[command(subcommand)]
-    command: AgentSessionCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum AgentSessionCommand {
-    Start(AgentSessionStartArgs),
-    List(AgentSessionListArgs),
-    Attach(AgentSessionAttachArgs),
-}
-
-#[derive(Debug, Args)]
-struct AgentSessionStartArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long, value_enum, default_value = "main")]
-    lane: AgentSessionLaneArg,
-    #[arg(long = "run")]
-    run_id: Option<String>,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct SessionArgs {
-    #[command(subcommand)]
-    command: SessionCommand,
-}
-
-#[derive(Debug, Args)]
-struct ProjectArgs {
-    #[command(subcommand)]
-    command: ProjectCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum ProjectCommandArgs {
-    #[command(about = "Read tracker state and Project health")]
-    State(ProjectStateArgs),
-    #[command(about = "Read one Project issue and linked PR evidence")]
-    Issue(ProjectIssueArgs),
-    #[command(about = "Inspect live issue readiness without mutating tracker state")]
-    Inspect(ProjectInspectArgs),
-    #[command(name = "set-state", about = "Set one issue Project status")]
-    SetState(SetStateArgs),
-    #[command(name = "link-pr", about = "Record pull request evidence for one issue")]
-    LinkPr(LinkPrArgs),
-    #[command(name = "add", about = "Add one GitHub issue to the configured Project")]
-    Add(AddToProjectArgs),
-    #[command(about = "Upsert the canonical issue workpad")]
-    Workpad(WorkpadArgs),
-    #[command(
-        name = "timeline-comment",
-        about = "Append a standalone issue timeline comment"
-    )]
-    TimelineComment(TimelineCommentArgs),
-}
-
-#[derive(Debug, Args)]
-struct ProjectInspectArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(help = "Issue identifier to inspect, for example #284")]
-    issue_ref: String,
-    #[arg(long, value_enum, help = "Optional lane context for readiness output")]
-    lane: Option<AgentSessionLaneArg>,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "write")]
-    _write: bool,
-}
-
-#[derive(Debug, Args)]
-struct TimelineCommentArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(value_name = "MARKDOWN_PATH")]
-    markdown_path: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct MergeArgs {
-    #[command(subcommand)]
-    command: MergeCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum MergeCommandArgs {
-    Claim(LaneClaimArgs),
-    Once(MergeOnceArgs),
-    Loop(MergeLoopArgs),
-}
-
-#[derive(Debug, Subcommand)]
-enum SessionCommand {
-    Start(SessionStartArgs),
-    List(AgentSessionListArgs),
-    Attach(AgentSessionAttachArgs),
-}
-
-#[derive(Debug, Args)]
-struct SessionStartArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long, value_enum)]
-    lane: AgentSessionLaneArg,
-    #[arg(long = "run")]
-    run_id: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct AgentSessionListArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-}
-
-#[derive(Debug, Args)]
-struct AgentSessionAttachArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    session: String,
-    #[arg(long)]
-    exec: bool,
-}
-
-#[derive(Debug, Args)]
-struct LaneSessionAliasArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct GateArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct SetStateArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    state: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct WorkpadArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    markdown_path: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct LinkPrArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    pr_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct CreateFollowUpArgs {
-    #[arg(long)]
-    workflow: PathBuf,
-    #[arg(long)]
-    title: String,
-    #[arg(long = "body-file")]
-    body_file: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct AddToProjectArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_id: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewFakeArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long, value_enum, default_value = "pass")]
-    outcome: CliFakeReviewOutcome,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewOnceArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewClaimArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    worker: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewClearClaimArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewEvidenceArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long = "evidence-file")]
-    evidence_file: PathBuf,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewRejectArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md")]
-    workflow_path: PathBuf,
-    issue_ref: String,
-    #[arg(long = "evidence-file")]
-    evidence_file: PathBuf,
-    #[arg(long = "target-state", default_value = "agent_review")]
-    target_state: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewFreshnessArgs {
-    #[arg(long = "issue")]
-    issue_ref: String,
-    #[arg(long = "prior-head")]
-    prior_head_sha: String,
-    #[arg(long = "current-head")]
-    current_head_sha: String,
-    #[arg(long = "prior-base")]
-    prior_base_sha: String,
-    #[arg(long = "current-base")]
-    current_base_sha: String,
-    #[arg(long = "changed-file")]
-    changed_files: Vec<String>,
-    #[arg(long = "stale-reason", value_enum)]
-    stale_reason: CliReviewStaleReason,
-    #[arg(long = "rework-class", value_enum)]
-    rework_class: CliReviewReworkClass,
-    #[arg(long = "patch-summary")]
-    patch_summary: Option<String>,
-}
-
-#[derive(Debug, Args)]
-struct ReviewLoopArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long)]
-    max_iterations: Option<usize>,
-    #[arg(long)]
-    once: bool,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "max-concurrent")]
-    max_concurrent: Option<usize>,
-    #[arg(long = "dry-run")]
-    _dry_run: bool,
-    #[arg(long = "fake-outcome", value_enum)]
-    fake_outcome: Option<CliFakeReviewOutcome>,
-}
-
-#[derive(Debug, Args)]
-struct ReviewStatusArgs {
-    #[arg(value_name = "path-to-WORKFLOW.md", default_value = "WORKFLOW.md")]
-    workflow_path: PathBuf,
-    #[arg(long = "issue", help = "Filter status to one issue, for example #313")]
-    issue_filter: Option<String>,
-    #[arg(
-        long = "recent",
-        default_value_t = DEFAULT_RECENT_REVIEW_JOBS,
-        help = "Number of recent completed or failed review jobs to show"
-    )]
-    recent_limit: usize,
-    #[arg(long, help = "Show more paths and anomaly details")]
-    verbose: bool,
-    #[arg(long, help = "Print the complete structured review status payload")]
-    json: bool,
-}
-
-#[derive(Debug, Args)]
-struct ReviewArgs {
-    #[command(subcommand)]
-    command: ReviewCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum ReviewCommandArgs {
-    Fake(ReviewFakeArgs),
-    Once(ReviewOnceArgs),
-    Claim(LaneClaimArgs),
-    Pass(ReviewEvidenceArgs),
-    Reject(ReviewRejectArgs),
-    Session(LaneSessionAliasArgs),
-    Freshness(ReviewFreshnessArgs),
-    Loop(ReviewLoopArgs),
-    Status(ReviewStatusArgs),
-}
-
-#[derive(Debug, Clone, Copy, ValueEnum)]
-enum CliReviewStaleReason {
-    MergeConflict,
-    BaseBranchUpdated,
-    ReviewOutdated,
-    Unknown,
-}
-
-impl From<CliReviewStaleReason> for ReviewStaleReason {
-    fn from(value: CliReviewStaleReason) -> Self {
-        match value {
-            CliReviewStaleReason::MergeConflict => ReviewStaleReason::MergeConflict,
-            CliReviewStaleReason::BaseBranchUpdated => ReviewStaleReason::BaseBranchUpdated,
-            CliReviewStaleReason::ReviewOutdated => ReviewStaleReason::ReviewOutdated,
-            CliReviewStaleReason::Unknown => ReviewStaleReason::Unknown,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, ValueEnum)]
-enum CliReviewReworkClass {
-    MechanicalConflictResolution,
-    BaseRefresh,
-    SemanticChange,
-    Unknown,
-}
-
-impl From<CliReviewReworkClass> for ReviewReworkClass {
-    fn from(value: CliReviewReworkClass) -> Self {
-        match value {
-            CliReviewReworkClass::MechanicalConflictResolution => {
-                ReviewReworkClass::MechanicalConflictResolution
-            }
-            CliReviewReworkClass::BaseRefresh => ReviewReworkClass::BaseRefresh,
-            CliReviewReworkClass::SemanticChange => ReviewReworkClass::SemanticChange,
-            CliReviewReworkClass::Unknown => ReviewReworkClass::Unknown,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, ValueEnum)]
-enum CliFakeReviewOutcome {
-    Pass,
-    Confirmed,
-    Failed,
-}
-
-impl From<CliFakeReviewOutcome> for FakeReviewOutcome {
-    fn from(value: CliFakeReviewOutcome) -> Self {
-        match value {
-            CliFakeReviewOutcome::Pass => FakeReviewOutcome::Pass,
-            CliFakeReviewOutcome::Confirmed => FakeReviewOutcome::ConfirmedFinding,
-            CliFakeReviewOutcome::Failed => FakeReviewOutcome::Failed,
-        }
-    }
-}
-
-#[derive(Debug, Args)]
-struct ForgeMarkdownArgs {
-    #[arg(long)]
-    body: Option<String>,
-    #[arg(long = "body-file", alias = "file")]
-    body_file: Option<PathBuf>,
-}
-
-#[derive(Debug, Args)]
-struct ForgeArgs {
-    #[command(subcommand)]
-    command: ForgeCommandArgs,
-}
-
-#[derive(Debug, Subcommand)]
-enum ForgeCommandArgs {
-    Create(ForgeCreateArgs),
-    Promote(ForgePromoteArgs),
-    Rework(ForgeReworkArgs),
-    Validate(ForgeValidateArgs),
-}
-
-#[derive(Debug, Args)]
-struct ForgeCreateArgs {
-    #[arg(long, default_value = "workflows/jade-symphony.md")]
-    workflow: PathBuf,
-    #[arg(long)]
-    title: String,
-    #[command(flatten)]
-    markdown: ForgeMarkdownArgs,
-    #[arg(long, value_enum, ignore_case = true, default_value_t = ForgeStatusArg::Todo)]
-    status: ForgeStatusArg,
-    #[arg(long)]
-    project: Option<String>,
-    #[arg(long = "project-field")]
-    project_fields: Vec<String>,
-    #[arg(long = "assignee")]
-    assignees: Vec<String>,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ForgePromoteArgs {
-    issue_ref: String,
-    #[arg(long, default_value = "workflows/jade-symphony.md")]
-    workflow: PathBuf,
-    #[arg(long)]
-    title: String,
-    #[command(flatten)]
-    markdown: ForgeMarkdownArgs,
-    #[command(flatten)]
-    promotion_note: PromotionNoteArgs,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct ForgeReworkArgs {
-    issue_ref: String,
-    #[arg(long, default_value = "workflows/jade-symphony.md")]
-    workflow: PathBuf,
-    #[arg(long)]
-    title: String,
-    #[command(flatten)]
-    markdown: ForgeMarkdownArgs,
-    #[arg(long = "evidence-file")]
-    evidence_file: PathBuf,
-    #[arg(long = "operator-confirmation")]
-    operator_confirmation: String,
-    #[arg(long)]
-    write: bool,
-    #[arg(long = "dry-run")]
-    dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct PromotionNoteArgs {
-    #[arg(long = "operator-confirmation")]
-    operator_confirmation: String,
-    #[arg(long = "decision", required = true)]
-    decisions: Vec<String>,
-    #[arg(long = "scope-change", required = true)]
-    scope_changes: Vec<String>,
-    #[arg(long = "dependency-context", required = true)]
-    dependencies_context: Vec<String>,
-    #[arg(long = "readback-summary")]
-    readback_summaries: Vec<String>,
-}
-
-#[derive(Debug, Args)]
-struct ForgeValidateArgs {
-    #[arg(long, default_value = "workflows/jade-symphony.md")]
-    workflow: PathBuf,
-    #[arg(long, value_enum, ignore_case = true)]
-    status: Option<ForgeStatusArg>,
-    #[arg(long)]
-    title: Option<String>,
-    #[command(flatten)]
-    markdown: ForgeMarkdownArgs,
-    #[arg(long = "issue")]
-    issue_ref: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
-enum ForgeStatusArg {
-    Backlog,
-    Todo,
-}
-
-impl ForgeStatusArg {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Backlog => "Backlog",
-            Self::Todo => "Todo",
-        }
-    }
-
-    fn normalized_state(self) -> &'static str {
-        match self {
-            Self::Backlog => "backlog",
-            Self::Todo => "todo",
-        }
-    }
-}
-
-fn run_loop_command(args: RunLoopArgs) -> Result<Command, String> {
-    if args.max_iterations == Some(0) || args.max_concurrent == Some(0) {
-        return Err(usage());
-    }
-    Ok(Command::RunLoop {
-        options: RunLoopOptions {
-            workflow_path: args.workflow_path,
-            max_iterations: args.max_iterations,
-            once: args.once,
-            write: args.write,
-            recover: loop_recover_enabled(args.write, args.recover, args.no_recover),
-            max_concurrent: args.max_concurrent,
-            display: args.display.into(),
-        },
-    })
-}
-
-fn merge_loop_command(args: MergeLoopArgs) -> Result<Command, String> {
-    if args.max_iterations == Some(0)
-        || args.max_concurrent == Some(0)
-        || (!args.once && args.max_iterations.is_none())
-    {
-        return Err(usage());
-    }
-    Ok(Command::MergeLoop {
-        options: MergeLoopOptions {
-            workflow_path: args.workflow_path,
-            max_iterations: args.max_iterations,
-            once: args.once,
-            write: args.write,
-            recover: loop_recover_enabled(args.write, args.recover, args.no_recover),
-            max_concurrent: args.max_concurrent,
-        },
-    })
-}
-
-fn loop_recover_enabled(write: bool, explicit_recover: bool, no_recover: bool) -> bool {
-    explicit_recover || (write && !no_recover)
-}
-
-fn command_from_project_args(command: ProjectCommandArgs) -> Result<Command, String> {
-    match command {
-        ProjectCommandArgs::State(args) => Ok(Command::ProjectState {
-            options: ProjectStateOptions {
-                workflow_path: args.workflow_path,
-                display: args.display.into(),
-            },
-        }),
-        ProjectCommandArgs::Issue(args) => Ok(Command::ProjectIssue {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            json: args.json,
-        }),
-        ProjectCommandArgs::Inspect(args) => Ok(Command::ProjectInspect {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            lane: args.lane,
-        }),
-        ProjectCommandArgs::SetState(args) => Ok(Command::SetState {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            state: args.state,
-            write: args.write,
-        }),
-        ProjectCommandArgs::LinkPr(args) => Ok(Command::LinkPr {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            pr_ref: args.pr_ref,
-            write: args.write,
-        }),
-        ProjectCommandArgs::Add(args) => Ok(Command::AddToProject {
-            workflow_path: args.workflow_path,
-            issue_id: args.issue_id,
-            write: args.write,
-        }),
-        ProjectCommandArgs::Workpad(args) => Ok(Command::Workpad {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            markdown_path: args.markdown_path,
-            write: args.write,
-        }),
-        ProjectCommandArgs::TimelineComment(args) => Ok(Command::TimelineComment {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            markdown_path: args.markdown_path,
-            write: args.write,
-        }),
-    }
-}
-
-fn command_from_merge_args(command: MergeCommandArgs) -> Result<Command, String> {
-    match command {
-        MergeCommandArgs::Claim(claim) => Ok(Command::LaneClaim {
-            workflow_path: claim.workflow_path,
-            issue_ref: claim.issue_ref,
-            lane: AgentSessionLaneArg::Merge,
-            worker: claim.worker,
-            source: claim.source,
-            write: claim.write,
-        }),
-        MergeCommandArgs::Once(args) => Ok(Command::MergeOnce {
-            workflow_path: args.workflow_path,
-            write: args.write,
-        }),
-        MergeCommandArgs::Loop(args) => merge_loop_command(args),
-    }
-}
-
-impl TryFrom<Cli> for Command {
-    type Error = String;
-
-    fn try_from(cli: Cli) -> Result<Self, Self::Error> {
-        let default_workflow = || PathBuf::from("WORKFLOW.md");
-        match cli.command {
-            None => Ok(Self::Plan {
-                workflow_path: cli.workflow_path.unwrap_or_else(default_workflow),
-                json: false,
-            }),
-            Some(command) => {
-                if cli.workflow_path.is_some() {
-                    return Err(usage());
-                }
-
-                match command {
-                    CliCommand::Plan(args) => Ok(Self::Plan {
-                        workflow_path: args.workflow_path,
-                        json: args.json,
-                    }),
-                    CliCommand::Validate(args) => Ok(Self::Validate {
-                        workflow_path: args.workflow_path,
-                    }),
-                    CliCommand::Doctor(args) => Ok(Self::Doctor {
-                        options: DoctorOptions {
-                            workflow_path: args.workflow_path,
-                            json: args.json,
-                            strict: args.strict,
-                            display: args.display.into(),
-                            interactive: args.interactive,
-                            auto_fix: args.auto_fix,
-                            write: args.write,
-                            stale_after_ms: args.stale_after_ms,
-                            action: args.action.map(|action| match action {
-                                DoctorSubcommandArgs::Repair(repair) => {
-                                    DoctorAction::Repair(DoctorRepairIssueOptions {
-                                        issue_ref: repair.issue_ref,
-                                        write: repair.write,
-                                        move_need_human_input: repair.move_need_human_input,
-                                        mark_pr_ready: repair.mark_pr_ready,
-                                        confirm_handoff_ready: repair.confirm_handoff_ready,
-                                    })
-                                }
-                            }),
-                        },
-                    }),
-                    CliCommand::DoctorRepairHumanReview(args) => {
-                        Ok(Self::DoctorRepairHumanReview {
-                            workflow_path: args.workflow_path,
-                            write: args.write,
-                        })
-                    }
-                    CliCommand::Skills(args) => match args.command {
-                        SkillsCommandArgs::Status(args) => Ok(Self::SkillsStatus {
-                            input: SkillStatusInput {
-                                workflow_path: args.workflow_path,
-                                suite_path: args.suite_path,
-                                codex_dir: args.codex_dir,
-                                gemini_dir: args.gemini_dir,
-                                require_gemini: args.require_gemini,
-                                session_skills: args.session_skills,
-                                session_skills_file: args.session_skills_file,
-                            },
-                            json: args.json,
-                        }),
-                    },
-                    CliCommand::Profiles(args) => Ok(Self::Profiles {
-                        workflow_path: args.workflow_path,
-                    }),
-                    CliCommand::Debug(args) => Ok(Self::Debug {
-                        workflow_path: args.workflow_path,
-                    }),
-                    CliCommand::Autopilot(args) => match args.command {
-                        AutopilotCommandArgs::Plan(args) => Ok(Self::AutopilotPlan {
-                            workflow_path: args.workflow_path,
-                            json: args.json,
-                        }),
-                    },
-                    CliCommand::Status(args) => match args.command {
-                        StatusCommandArgs::Show(show) => Ok(Self::Plan {
-                            workflow_path: show.workflow_path,
-                            json: show.json,
-                        }),
-                        StatusCommandArgs::Serve(serve) => Ok(Self::StatusApi {
-                            workflow_path: serve.workflow_path,
-                            bind: serve.bind,
-                            once: serve.once,
-                        }),
-                    },
-                    CliCommand::Clean(args) => match args.command {
-                        CleanCommand::Plan(plan) => Ok(Self::CleanPlan {
-                            workflow_path: plan.workflow_path,
-                        }),
-                        CleanCommand::Audit(audit) => Ok(Self::CleanAudit {
-                            workflow_path: audit.workflow_path,
-                        }),
-                    },
-                    CliCommand::Workspace(args) => match args.command {
-                        WorkspaceCommandArgs::List(list) => Ok(Self::WorkspaceList {
-                            workflow_path: list.workflow_path,
-                        }),
-                        WorkspaceCommandArgs::Show(show) => Ok(Self::WorkspaceShow {
-                            workflow_path: show.workflow_path,
-                            issue_ref: show.issue_ref,
-                        }),
-                        WorkspaceCommandArgs::Adopt(adopt) => Ok(Self::WorkspaceAdopt {
-                            workflow_path: adopt.workflow_path,
-                            issue_ref: adopt.issue_ref,
-                            path: adopt.path,
-                            write: adopt.write,
-                        }),
-                        WorkspaceCommandArgs::Ensure(ensure) => Ok(Self::WorkspaceEnsure {
-                            workflow_path: ensure.workflow_path,
-                            issue_ref: ensure.issue_ref,
-                            pr_ref: ensure.pr_ref,
-                            branch: ensure.branch,
-                            write: ensure.write,
-                        }),
-                    },
-                    CliCommand::Project(args) => command_from_project_args(args.command),
-                    CliCommand::Main(args) => lane_command(AgentSessionLaneArg::Main, args),
-                    CliCommand::Merge(args) => command_from_merge_args(args.command),
-                    CliCommand::Session(args) => match args.command {
-                        SessionCommand::Start(start) => Ok(Self::SessionStart {
-                            workflow_path: start.workflow_path,
-                            issue_ref: start.issue_ref,
-                            lane: start.lane,
-                            run_id: start.run_id,
-                            write: start.write,
-                        }),
-                        SessionCommand::List(list) => Ok(Self::SessionList {
-                            workflow_path: list.workflow_path,
-                        }),
-                        SessionCommand::Attach(attach) => Ok(Self::SessionAttach {
-                            workflow_path: attach.workflow_path,
-                            session: attach.session,
-                            exec: attach.exec,
-                        }),
-                    },
-                    CliCommand::CreateFollowUp(args) => Ok(Self::CreateFollowUp {
-                        workflow_path: args.workflow,
-                        title: args.title,
-                        body_path: args.body_file,
-                        write: args.write,
-                    }),
-                    CliCommand::Review(args) => command_from_review_args(args.command),
-                    CliCommand::Forge(args) => match args.command {
-                        ForgeCommandArgs::Create(args) => Ok(Self::ForgeCreate {
-                            workflow_path: args.workflow,
-                            title: args.title,
-                            markdown: read_forge_markdown_arg(args.markdown)?,
-                            status: args.status,
-                            project: args.project,
-                            project_fields: parse_project_field_assignments(args.project_fields)?,
-                            assignees: args.assignees,
-                            write: args.write,
-                            dry_run: args.dry_run,
-                        }),
-                        ForgeCommandArgs::Promote(args) => Ok(Self::ForgePromote {
-                            workflow_path: args.workflow,
-                            issue_ref: args.issue_ref,
-                            title: args.title,
-                            markdown: read_forge_markdown_arg(args.markdown)?,
-                            promotion_note: promotion_note_input(args.promotion_note)?,
-                            write: args.write,
-                            dry_run: args.dry_run,
-                        }),
-                        ForgeCommandArgs::Rework(args) => Ok(Self::ForgeRework {
-                            options: ForgeReworkOptions {
-                                workflow_path: args.workflow,
-                                issue_ref: args.issue_ref,
-                                title: args.title,
-                                markdown: read_forge_markdown_arg(args.markdown)?,
-                                evidence: read_required_file(args.evidence_file)?,
-                                operator_confirmation: args.operator_confirmation,
-                                write: args.write,
-                                dry_run: args.dry_run,
-                            },
-                        }),
-                        ForgeCommandArgs::Validate(args) => {
-                            if let Some(issue_ref) = args.issue_ref {
-                                Ok(Self::ForgeValidate {
-                                    workflow_path: args.workflow,
-                                    status: args.status,
-                                    title: args.title.unwrap_or_default(),
-                                    markdown: read_optional_forge_markdown_arg(args.markdown)?,
-                                    issue_ref: Some(issue_ref),
-                                })
-                            } else {
-                                Ok(Self::ForgeValidate {
-                                    workflow_path: args.workflow,
-                                    status: args.status,
-                                    title: args.title.ok_or(
-                                        "forge validate requires --title when --issue is not used",
-                                    )?,
-                                    markdown: read_forge_markdown_arg(args.markdown)?,
-                                    issue_ref: None,
-                                })
-                            }
-                        }
-                    },
-                    CliCommand::Run => {
-                        Err("`jade-symphony run` is reserved for future all-lane orchestration and is not implemented yet".into())
-                    }
-                    CliCommand::Upgrade => {
-                        Err("`jade-symphony upgrade` is reserved for future Jade Symphony binary and skill upgrades and is not implemented yet".into())
-                    }
-                }
-            }
-        }
-    }
-}
-
 fn gate_workpad(issue: &TrackerIssue, decision: &GateDecision) -> String {
     let mut lines = vec![
         "## Jade Symphony Workpad".to_string(),
@@ -15740,161 +14306,6 @@ fn gate_target_state(decision: &GateDecision) -> &'static str {
     }
 }
 
-fn read_forge_markdown_arg(args: ForgeMarkdownArgs) -> Result<String, String> {
-    match (args.body, args.body_file) {
-        (Some(value), None) => Ok(value),
-        (None, Some(path)) => std::fs::read_to_string(&path)
-            .map_err(|error| format!("failed to read {}: {error}", path.display())),
-        _ => Err(usage()),
-    }
-}
-
-fn read_optional_forge_markdown_arg(args: ForgeMarkdownArgs) -> Result<String, String> {
-    match (args.body, args.body_file) {
-        (Some(value), None) => Ok(value),
-        (None, Some(path)) => std::fs::read_to_string(&path)
-            .map_err(|error| format!("failed to read {}: {error}", path.display())),
-        (None, None) => Ok(String::new()),
-        (Some(_), Some(_)) => Err(usage()),
-    }
-}
-
-fn read_required_file(path: PathBuf) -> Result<String, String> {
-    std::fs::read_to_string(&path)
-        .map_err(|error| format!("failed to read {}: {error}", path.display()))
-}
-
-fn command_from_review_args(command: ReviewCommandArgs) -> Result<Command, String> {
-    match command {
-        ReviewCommandArgs::Fake(args) => Ok(Command::ReviewFake {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            outcome: args.outcome.into(),
-            write: args.write,
-        }),
-        ReviewCommandArgs::Once(args) => Ok(Command::ReviewOnce {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            write: args.write,
-        }),
-        ReviewCommandArgs::Claim(args) => Ok(Command::LaneClaim {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            lane: AgentSessionLaneArg::Review,
-            worker: args.worker,
-            source: args.source,
-            write: args.write,
-        }),
-        ReviewCommandArgs::Pass(args) => Ok(Command::ReviewPass {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            evidence: read_required_file(args.evidence_file)?,
-            write: args.write,
-        }),
-        ReviewCommandArgs::Reject(args) => Ok(Command::ReviewReject {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            evidence: read_required_file(args.evidence_file)?,
-            target_state: args.target_state,
-            write: args.write,
-        }),
-        ReviewCommandArgs::Session(args) => Ok(Command::ReviewSession {
-            workflow_path: args.workflow_path,
-            issue_ref: args.issue_ref,
-            write: args.write,
-        }),
-        ReviewCommandArgs::Freshness(args) => Ok(Command::ReviewFreshness {
-            input: ReviewFreshnessInput {
-                issue_ref: args.issue_ref,
-                prior_head_sha: args.prior_head_sha,
-                current_head_sha: args.current_head_sha,
-                prior_base_sha: args.prior_base_sha,
-                current_base_sha: args.current_base_sha,
-                changed_files: args.changed_files,
-                stale_reason: args.stale_reason.into(),
-                rework_class: args.rework_class.into(),
-                patch_summary: args.patch_summary,
-            },
-        }),
-        ReviewCommandArgs::Loop(args) => {
-            if args.max_iterations == Some(0) || args.max_concurrent == Some(0) {
-                return Err(usage());
-            }
-            Ok(Command::ReviewLoop {
-                options: ReviewLoopOptions {
-                    workflow_path: args.workflow_path,
-                    max_iterations: args.max_iterations,
-                    once: args.once,
-                    write: args.write,
-                    fake_outcome: args.fake_outcome.map(Into::into),
-                    max_concurrent: args.max_concurrent,
-                },
-            })
-        }
-        ReviewCommandArgs::Status(args) => {
-            if args.recent_limit == 0 {
-                return Err(usage());
-            }
-            Ok(Command::ReviewStatus {
-                options: ReviewStatusCliOptions {
-                    workflow_path: args.workflow_path,
-                    issue_filter: args.issue_filter,
-                    recent_limit: args.recent_limit,
-                    verbose: args.verbose,
-                    json: args.json,
-                },
-            })
-        }
-    }
-}
-
-fn parse_project_field_assignments(
-    values: Vec<String>,
-) -> Result<Vec<ProjectFieldAssignment>, String> {
-    values
-        .into_iter()
-        .map(|value| ProjectFieldAssignment::parse(&value).map_err(|error| error.to_string()))
-        .collect()
-}
-
-fn promotion_note_input(args: PromotionNoteArgs) -> Result<PromotionNoteInput, String> {
-    fn clean_nonempty(value: String, field: &str) -> Result<String, String> {
-        let trimmed = value.trim().to_string();
-        if trimmed.is_empty() {
-            Err(format!("forge promote requires non-empty {field}"))
-        } else {
-            Ok(trimmed)
-        }
-    }
-
-    fn clean_many(values: Vec<String>, field: &str) -> Result<Vec<String>, String> {
-        let cleaned = values
-            .into_iter()
-            .map(|value| clean_nonempty(value, field))
-            .collect::<Result<Vec<_>, _>>()?;
-        if cleaned.is_empty() {
-            Err(format!("forge promote requires at least one {field}"))
-        } else {
-            Ok(cleaned)
-        }
-    }
-
-    Ok(PromotionNoteInput {
-        operator_confirmation: clean_nonempty(
-            args.operator_confirmation,
-            "--operator-confirmation",
-        )?,
-        decisions: clean_many(args.decisions, "--decision")?,
-        scope_changes: clean_many(args.scope_changes, "--scope-change")?,
-        dependencies_context: clean_many(args.dependencies_context, "--dependency-context")?,
-        readback_summaries: args
-            .readback_summaries
-            .into_iter()
-            .map(|value| clean_nonempty(value, "--readback-summary"))
-            .collect::<Result<Vec<_>, _>>()?,
-    })
-}
-
 fn print_forge_validation(report: &ForgeValidationReport) {
     let categories = forge_missing_categories(report);
     println!("title={}", report.title);
@@ -15926,51 +14337,6 @@ fn missing_category_value(values: &[String]) -> String {
     } else {
         values.join(", ")
     }
-}
-
-fn usage() -> String {
-    [
-        "OpenAI Symphony-style orchestration harness with Jade Symphony extensions",
-        "",
-        "Usage: jade-symphony [path-to-WORKFLOW.md] [COMMAND]",
-        "",
-        "Human / Operator operations:",
-        "  plan                        Render the dispatch/status plan",
-        "  validate                    Validate workflow loading and configuration",
-        "  doctor                      Audit Project, workflow, and runtime invariants",
-        "  skills                      Inspect per-repo skill readiness",
-        "  status                      Show or serve runtime status snapshots",
-        "  clean                       Plan or audit artifact cleanup",
-        "  profiles                    List execution profiles",
-        "  debug                       Render a combined operator debug report",
-        "",
-        "Project / Agent internals:",
-        "  project                     Read or mutate Project facts through grouped subcommands",
-        "  workspace                   Discover and record per-issue git worktrees",
-        "  session                     Start, list, or attach supervised lane sessions",
-        "",
-        "Lane orchestration:",
-        "  autopilot                   Read-only all-lane planning preflight",
-        "  main                        Main Agent claim, once, and loop commands",
-        "  review                      Review Agent claim, pass/reject, session, freshness, and loop commands",
-        "  merge                       Merging Agent claim, once, and loop commands",
-        "  create-follow-up            Create an operator follow-up issue",
-        "",
-        "Issue Forge:",
-        "  forge                       Validate, create, or promote issue contracts",
-        "",
-        "Reserved lifecycle topology:",
-        "  run                         Reserved for future all-lane automatic orchestration",
-        "  upgrade                     Reserved for future Jade Symphony binary and skill upgrades",
-        "",
-        "Arguments:",
-        "  [path-to-WORKFLOW.md]",
-        "",
-        "Options:",
-        "  -h, --help                  Print help",
-        "",
-    ]
-    .join("\n")
 }
 
 #[cfg(test)]
